+++ src/OpenTelemetry.Api/Context/Propagation/HttpParseResult.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="DistributedContextBinarySerializer.cs" company="OpenTelemetry Authors">
+﻿// <copyright file="HttpParseResult.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,22 +16,22 @@
 
 namespace OpenTelemetry.Context.Propagation
 {
-    internal sealed class DistributedContextBinarySerializer : DistributedContextBinarySerializerBase
+    // Adoptation of code from https://github.com/aspnet/HttpAbstractions/blob/07d115400e4f8c7a66ba239f230805f03a14ee3d/src/Microsoft.Net.Http.Headers/HttpParseResult.cs
+    internal enum HttpParseResult
     {
-        internal DistributedContextBinarySerializer()
-        {
-        }
+        /// <summary>
+        /// Parsed successfully.
+        /// </summary>
+        Parsed,
 
-        /// <inheritdoc/>
-        public override byte[] ToByteArray(DistributedContext context)
-        {
-            return SerializationUtils.SerializeBinary(context);
-        }
+        /// <summary>
+        /// Was not parsed.
+        /// </summary>
+        NotParsed,
 
-        /// <inheritdoc/>
-        public override DistributedContext FromByteArray(byte[] bytes)
-        {
-            return SerializationUtils.DeserializeBinary(bytes);
-        }
+        /// <summary>
+        /// Invalid format.
+        /// </summary>
+        InvalidFormat,
     }
 }