+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Changed the default propagation to support W3C Baggage
+  ([#1048](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1048))
+  * The default ITextFormat is now `CompositePropagator(TraceContextFormat,
+    BaggageFormat)`. Outgoing Http request will now send Baggage using the [W3C
+    Baggage](https://github.com/w3c/baggage/blob/master/baggage/HTTP_HEADER_FORMAT.md)
+    header. Previously Baggage was sent using the `Correlation-Context` header,
+    which is now outdated.
 * Removed `AddHttpInstrumentation` and `AddHttpWebRequestInstrumentation` (.NET
   Framework) `TracerProviderBuilderExtensions`. `AddHttpClientInstrumentation`
   will now register `HttpClient` instrumentation on .NET Core and `HttpClient` +